+++ src/ast/irbuilderbpf.h
@@ -129,15 +129,7 @@ public:
   Value *CreateStrcontains(Value *val1,
                            uint64_t str1_size,
                            Value *val2,
-                           uint64_t str2_size,
-                           bool inverse);
-  Value *CreateIntegerArrayCmpUnrolled(Value *ctx,
-                                       Value *val1,
-                                       Value *val2,
-                                       const SizedType &val1_type,
-                                       const SizedType &val2_type,
-                                       const bool inverse,
-                                       const location &loc);
+                           uint64_t str2_size);
   Value *CreateIntegerArrayCmp(Value *ctx,
                                Value *val1,
                                Value *val2,