+++ tests/codegen/builtin_pid_tid.cpp
@@ -11,12 +11,11 @@ TEST(codegen, builtin_pid_tid)
 
 TEST(codegen, builtin_pid_tid_namespace)
 {
-  MockBPFtrace bpftrace;
-  bpftrace.feature_ = std::make_unique<MockBPFfeature>(true);
-  bpftrace.mock_in_init_pid_ns = false;
-  bpftrace.helper_check_level_ = 0;
+  auto bpftrace = get_mock_bpftrace();
+  bpftrace->mock_in_init_pid_ns = false;
+  bpftrace->helper_check_level_ = 0;
 
-  test(bpftrace, "kprobe:f { @x = pid; @y = tid }", NAME);
+  test(*bpftrace, "kprobe:f { @x = pid; @y = tid }", NAME);
 }
 
 } // namespace codegen